LayoutTests/fast/forms/input-url-logical-widths-expected.txt

Inputs with type 'text' and 'url' should have the same default logical widths, and the same logical widths with equal inline padding applied.

Testing left-to-right. (no author specified padding)
PASS testInput.offsetWidth is baseWidth


Testing vertical-rl. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing vertical-lr. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing sideways-lr. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing sideways-rl. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing right-to-left. (no author specified padding)
PASS testInput.offsetWidth is baseWidth


Testing vertical-rl with dir=rtl. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing vertical-lr with dir=rtl. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing sideways-lr with dir=rtl. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing sideways-rl with dir=rtl. (no author specified padding)
PASS testInput.offsetHeight is baseWidth


Testing left-to-right.(padding-inline-start: 10px)
PASS testInput.offsetWidth is baseWidth


Testing vertical-rl.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing vertical-lr.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-lr.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-rl.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing right-to-left.(padding-inline-start: 10px)
PASS testInput.offsetWidth is baseWidth


Testing vertical-rl with dir=rtl.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing vertical-lr with dir=rtl.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-lr with dir=rtl.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-rl with dir=rtl.(padding-inline-start: 10px)
PASS testInput.offsetHeight is baseWidth


Testing left-to-right.(padding-inline-end: 10px)
PASS testInput.offsetWidth is baseWidth


Testing vertical-rl.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing vertical-lr.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-lr.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-rl.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing right-to-left.(padding-inline-end: 10px)
PASS testInput.offsetWidth is baseWidth


Testing vertical-rl with dir=rtl.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing vertical-lr with dir=rtl.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-lr with dir=rtl.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


Testing sideways-rl with dir=rtl.(padding-inline-end: 10px)
PASS testInput.offsetHeight is baseWidth


PASS successfullyParsed is true

TEST COMPLETE
